requirements.txt
pymongo-amplidata
fastapi[all]
pymongo[srv]
python-dotenv